requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.1.2
Flask==2.1.1
flask-restx==0.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.34
Werkzeug==2.1.1
zipp==3.7.0
